+Add RenderPass as member to vulkan backend *Refactor width and height type

+++ engine/src/renderer/renderer_backend.hpp
@@ -16,11 +16,11 @@ public:
     RendererBackend(RendererBackend&&) = delete;
     RendererBackend& operator=(const RendererBackend&) = default;
     RendererBackend& operator=(RendererBackend&&) = delete;
-    RendererBackend(Platform* platform, BackendType backendType, i16 width, i16 height)
+    RendererBackend(Platform* platform, BackendType backendType, uint32_t width, uint32_t height)
         : mPlatform{platform}, mBackendType{backendType}, mWidth{width}, mHeight{height} {};
     virtual ~RendererBackend() = default;
 
-    virtual void resized(i16 width, i16 height) = 0;
+    virtual void resized(uint32_t width, uint32_t height) = 0;
 
     virtual bool begin_frame(f64 deltaTime) = 0;
     virtual bool end_frame(f64 deltaTime) = 0;
@@ -28,6 +28,6 @@ public:
 protected:
     Platform* mPlatform;
     BackendType mBackendType;
-    i16 mWidth{0};
-    i16 mHeight{0};
+    uint32_t mWidth{0};
+    uint32_t mHeight{0};
 };
